(#3910) Default to no cache for nodes on master
The default cache terminus for nodes on the puppet master was yaml. If you are
using an ENC, there is no decent way to invalidate this cache. It does not
seem useful or necessary (the master only queries the node once per agent
check-in) so we're leaving it off by default.