(Maint) Mark failing tests as pending on Windows
A previous commit removed the fails_on_windows tag but did not add the
conditional pending block. Additional changes were made to faces
documentation that do not work on Windows.

This commit just adds a conditional pending block as was done in a
previous commit, e.g. for the mount type.