Maint: Escape command in regexp
On 64-bit Windows, the directory for 32-bit applications contains
parentheses: C:/Program Files (x86). The test was using this in a regexp
comparison and failing because the parentheses are metacharcters that
need to be escaped.

This failure was introduced when the `fails_on_windows` exclude filter
was removed (as `Puppet::Util.which` now works on Windows).