(maint) Get provider specs working on windows
Define helper to provider an existing command for windows because "echo"
is a shell interpreted command, not a proper executable.

As a side note, a helper method was the only way this actually functions
because of the evaluation context of the confine scenarios hash. Let,
before(:each), and instance variable all fail.