(#14137) resource_type serialization naming conventions
This commit attempts to bring the serialized representation
of resource_type objects into better alignment with naming
conventions that we've established in our documentation
and elsewhere.  In particular:

* "resource_supertype" => "kind"
* "hostclass" => "class"
* "definition" => "defined_type"
* "arguments" => "parameters"

These changes only affect the serialization (#to_pson/#from_pson).
The internal representation still uses the old terminology.
Ideally we'd just change the internals to conform to the
naming conventions, but that is a larger changeset than we
can reasonably tackle right now.