Fix over-stubbing in tests revealed by agent changes.
Various tests of the agent stubbed `File.read` globally, which means that they
could pass if and only if there was only one read ever made by the agent in
the execution of "apply".

This was, in fact, true until recent changes to ensure the ENC could correctly
override the environment that the agent ran in - which, in turn, reveals the
coupling of these tests to the implementation.

This corrects those tests by using a real temporary file rather than stubbing
out all file reads in a shot.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>