(#14173) Enforce that filebucket paths must be absolute
If you passed the filebucket a relative path, perhaps by accidentally quoting
'false', it would happily accept that and create a directory relative to the
current working directory to store content in.

This enforces that the path must be absolute, as well as failing cleanly if
multiple paths are passed.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>