Add default_file_terminus setting
This setting is used to determine the default source of files accessed
with a puppet: url and no server specified, e.g. puppet:///path/to/file.
Before now, this was done by checking if Puppet[:name] was `apply`. Since
Puppet[:name] is removed and we don't want application specific logic spread
throughout the code, we are adding a setting specific to this particular
behavior. The `apply` application sets a default value of `file_terminus`,
which retains the previous behavior of `apply`. Otherwise the default value is
`rest`, causing Puppet[:server] to be used to retrieve the file.