(#11004) Fix regex for pkg provider line matching
The pkg provider uses the output of `pkg list` to check if a package is
present on the system. On solaris 10 the output showed a dedicated
`status` field to reflect `installed` or `known`. On Solaris 11 this
field is not present anymore, so it was marked as an optional field in
the regex that is used to parse the `pkg list` output. The field
delimiter (\s+) however were notaround not marked as optional so while

  library/python-2/pycurl-26 7.19.0.1-0.175.0.0.0.2.537  i--

worked

  library/python-2/pycurl-26 7.19.0.1-0.175.0.0.0.2.537 i--

did not (notice there is only one whitespace between version an flags).

Modify the regex so the second example is also interpreted correctly.