(#14200) Fix help face
This commit does the following:

1. Fix the help face so that--in the case where there exists both an
App and a Face by the same name, and the App is not just a stub for the
Face--help gives preference to the App instead of the Face.  The help
face (as a command line tool) is mostly going to be used to help
users navigate the other command line tools, so, e.g.,
"puppet help resource" should show help for the same tool that you
will get by running "puppet resource".  This was not the case
prior to this commit.
2. Get rid of the separation between faces and applications in the
help screen... making this distinction seems more confusing than
helpful to the users, and exposes implementation details that
users should not need to be concerned with.
3. Change all class names for Applications to conform to the
convention of "foo_bar" file names mapping to "FooBar" class names.