Save face loader state in a different, more robust way.
The previous version mostly worked, but wound up showing an odd failure mode
under Ruby 1.9.3 where it didn't seem to restore state.  My best guess is that
something caused the older value to be closed over...

In any case this is an equally horrible, but more effective, way to resolve
the same problem.  I don't mind *like* either of them, but they work.

Really, the whole thing should be rethought - it may not actually be worth the
level of pain required to test the loader in the unit tests.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>