(#14136) Emit debug/warnings when Augeas provider fails to load files
The messages and information from /augeas//error are now printed to debug when
the provider opens Augeas, allowing file parse errors to be found.

The provider can also restrict the files that are loaded (from the default of
everything).  In these cases, the provider now emits a warning and then prints
all the detailed information to debug.  Since a few parse errors are common, no
warning is given when these optimisations aren't in use.