(#14288) gem provider should tolerate bad lines.
Previously, when the output of the gem command contained a line that didn't
match it was returned as `nil` - but the rest of the provider didn't handle
that value at all!

That led to any mismatch causing various failures to dereference `nil` as a
hash, and general failure.

This fixes this to behave significantly better by just omitting that entry in
the result, returning only the expected content.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>