(#14297) Handle upstart services better
The previous changes to the upstart provider didn't take into account services
that may have upstart jobs in /etc/init with no corresponding symlink to
upstart-job in /etc/init.d. This fix adds /etc/init/$service.conf to the search
path.

This commit also collapses some repeated code into a tighter loop for easier reading
and removes the debian provider as the default for ubuntu, as there were two
defaults for ubuntu. It also filters init.d scripts out of upstart jobs, so
service won't be called against them.

In order to allow upstart to use debian as its parent, this commit adds methods
for enabled?, enable and disable. Without this fix, using a debian style init
script on ubuntu requires manually specifying the debian provider be used.
With this commit, the upstart provider can be the default for ubuntu and
still fail up to its parent, debian.