(#14297) Update init and upstart service provider specs
As the service provider now looks in both /etc/init.d and /etc/init, the spec
test needs to reflect that change. This commit adds /etc/init to the expected
provider.defpath for the default case (not FreeBSD, HP-UX, or Archlinux).

This commit also updates existing rspec tests for the upstart provider. It adds
tests for the enabled?, enable, and disable methods that are now present in the
upstart service provider. It also adds some whitespace cleanup.