(#14297) Restore defpath, reduce duplicate code
This commit restores the defpath and moves that defpath change into the upstart
provider. It duplicates the search method from the init provider in the upstart
provider, but adds the .conf suffix to the upstart search. It also collapses
some repeated code in both the init and upstart providers into a simple loop.