Fix acceptance test assumption about exec logging.
The acceptance test for overriding the agent name fact was broken, in that it
assumed that `exec { "echo foo": }` would always write `foo` to the output
channel.

This isn't true - on success, that output was simply hidden.  By requesting
that output be logged we can ensure that message will show up and fix the
problem.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>