(#14297) Update upstart provider to handle multiline start stanzas
Upstart supports multiline 'start on' stanzas that prevent using a simple regex
to enable and disable upstart services. This commit updates the provider to do
some paren matching and comment/uncomment either the 'start on' with any
continuing lines appropriately. It comments lines until the net number of
opening/closing parens returns to 0.