Move rubygems require to executable entry points to Puppet
We were trying to load rubygems in lib/puppet.rb so that we could load other
libraries installed as gems, but this isn't anywhere near the first file
loaded. By moving this into bin/puppet and other executable entry points, we
ensure that rubygems is loaded as early as possible.