Cleanup of the Forge code
Removed some unreachable code that occured after an exception being
thrown. Pushed the construction of the Net::HTTP::Get object into the
repository so that it doesn't leak as much of its implementation out to
its clients.

Refactored the tests so that they include as many nested
contexts and pulled out helper methods to make the tests be more
explanatory.