(#3669) Disable SRV record use by default.
This is a nice feature to have, but it is not widely used.  It also triggers
some odd bugs in some common (embedded device) DNS recursive resolvers, which
simply ignore the question and lead to a long timeout waiting for the response
to come back.

Rather than ship this opt-out, it is safer to ship opt-in, and to allow people
to turn on the facility when they have configured it on their network.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>