(#14514) Check for the presence of hiera config
Puppet initializes Hiera using a configuration Hash instead of a
configuration file. Puppet now checks for the existence of the hiera
configuration file and uses an empty configuration and logs a warning
if missing.

We also ensure Hiera uses the Puppet logger by overriding the
configuration hash to set the logger explicitly.

This patch includes updated specs.