(#14715) Use gtar or Solaris tar properly with 'puppet module install' on Solaris
Previously we were assuming there was a usable GNU tar in the path for Solaris,
but Solaris provides its own tar implementation which works quite differently.

This patch changes the behaviour of the unpacker code during module
installation to use the 'gtar' implementation found on most Solaris 10 systems,
and later part of core with Solaris 11.

If 'gtar' is not found the behaviour is to fall back to Solaris tar instead,
using a gzip pipe to tar to get around the lack of decompression capabilities
in that implementation.

This also adds some unit tests for puppet/module_tool/applications/unpacker.