(Maint) Close filehandles after read in upstart
Not closing the filehandle after reading leaves a File object floating
around with an open handle. On unix systems this is ok, but on Windows
this causes a problem for the subsequent move of the new file into
place.

This changes to reading of a file to ensure that the file handle is
closed after we finish reading.