maint: Make it easier to test the service provider
Before the change the directories to search for initscripts (e.g.
`/etc/init.d`, `/sbin/init.d`) were stored in a class variable that was
calculated at load time. This way stubbing `operatingsystem` in test
cases did not change the value of `@defpath`.

Turning the class variable in a class method. This way stubbing
operatingsystem does change the result when querying defpath. The fact
lookup should be relativly cheap so we do not have to cache the
value in a class variable.