(#14615) Fix provider spec tests to work on non gentoo systems
The gentoo service provider tests made the assumption that /etc/init.d
is always the directory where the service provider will look for init
scripts while in fact the path depends on the operatingsystem.

The service provider spec also made the assumption that the sshd init
script is present on the target machine.

Stub the operatingsystem fact to prevent failures.
Stub the stat-call so the sshd script is always considered present.