(maint) Add mailmap to identify contributors
Without this patch it can be difficult to track a single contributor
over time.  This is because we switched version control systems early in
the project and individual contributors have used different email
addresses over time.

This patch partially fixes the problem by mapping some contributors to
their current identities.