(Maint) Clarify what certs are being created in ca tests
The make_certs method that was part of the test was not clear at the
point of call what the certs were that it was creating. It had two types
(requests and certificates) that it could make and most callers only
wanted one to be actaully made. This splits make_certs in two and
provides a clearer name for each part.